+++ packages/ragbits-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Add Prometheus & Grafana Monitoring for LLMs Using OpenTelemetry
 
+- Add token id lookup for LLMs (#513)
 - Allow custom source and element classes to be auto-imported (#343)
 - Make HuggingFaceSource row optional (#529)
 - Allow Source to be constructed from config (#529)